+++ lib/ui/pages/auth/join_page/widgets/login_move_button.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+import 'package:team_project/_core/constants/move.dart';
+import 'package:team_project/ui/widgets/custom_text_button.dart';
+
+class LoginMoveButton extends StatelessWidget {
+  const LoginMoveButton({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return CustomTextButton(
+      "로그인 페이지로 이동",
+      () {
+        Navigator.pushNamed(context, Move.loginPage);
+      },
+    );
+  }
+}
